% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmwg.R, R/tmp.R
\name{extract_pmwgs}
\alias{extract_pmwgs}
\title{Extract samples from a pmwgs object (from pmwg package)}
\usage{
extract_pmwgs(x)

extract_pmwgs(x)
}
\arguments{
\item{x}{The pmwgs object passed from is2 generic function}
}
\value{
A list with various samples, possibly reshaped

A list with various samples, possibly reshaped
}
\description{
Currently in rough form, must be from sample stage, too much hard coded.

Currently in rough form, must be from sample stage, too much hard coded.
}
