% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{get_logp}
\alias{get_logp}
\title{Calculate the log probability of the data given the xxx}
\usage{
get_logp(prop_theta, sampler, n_particles, subj_est, group_dist)
}
\arguments{
\item{prop_theta}{- the proposals for theta_mu, the group level model
parameter estimates}

\item{sampler}{- the object which holds the information about the sampling
process, including log likelihood function, data, n_subjects etc}

\item{n_particles}{- the number of particles to draw for each importance
sample.}

\item{subj_est}{- A named list containing the mean `mu` vector and covariance
matrices `sigma` of each subjects random effects as well as the number of
parameters used for the random effect estimates.}

\item{group_dist}{- The specific calculation for the log density for the
group distribution}
}
\value{
Something or ther - not sure yet.
}
\description{
Given th list of X parameters do some calucltions and return stuff
}
