% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{get_proposals}
\alias{get_proposals}
\title{Gets proposal mvnormal samples based on mix of gaussians}
\usage{
get_proposals(parvector, n_isamples, prop_args)
}
\arguments{
\item{parvector}{A 2D array with rows = sampler iterations and columns being
the model parameter estimates and other samples necessary.}

\item{n_isamples}{The number of importance samples to generate}

\item{prop_args}{The extra arguments passed from the is2 function}
}
\value{
An array with n_isamples proposal importance sample distributions
  drawn from the mixture of gaussians using rmvnorm.
}
\description{
Currently assumes that the mix of gaussians was generated with k = 2
}
