% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is2-package.R
\docType{package}
\name{is2-package}
\alias{is2-package}
\alias{_PACKAGE}
\title{is2: Importance Sampling Squared}
\description{
The is2 package provides a general purpose implementation of the
marginal likelihood calculation methods outlined in Tran et al. (2020)
\doi{10.3758/s13428-020-01348-w}. This package will work with pmwg sampler
objects, and can be also be used with the output of other MCMC samplers.
}
\section{Documentation}{

The documentation found at \url{https://newcastlecl.github.io/is2/} and
contains background information and motivation for the approach used in
this package and several detailed examples of the package in action.
}

\section{User input}{

The user is expected to provide a structured object with MCMC samples, a
log likelihood function. The initial version works well with a pmwg sampler
object.
}

\references{
Tran, M. N., Scharth, M., Gunawan, D., Kohn, R., Brown, S. D., &
Hawkins, G. E. (2020). Robustly estimating the marginal likelihood for
cognitive models via importance sampling.
\emph{Behavior Research Methods, 1-18.}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/newcastlecl/is2}
  \item Report bugs at \url{https://github.com/newcastlecl/is2/issues}
}

}
\author{
\strong{Maintainer}: Gavin Cooper \email{gavin@gavincooper.net} (Package creator and maintainer) [translator]

Authors:
\itemize{
  \item Reilly Innes \email{Reilly.Innes@uon.edu.au}
  \item Caroline Kuhne \email{caroline.kuhne@newcastle.edu.au}
  \item Jon-Paul Cavallaro \email{jon-paul.cavallaro@uon.edu.au}
  \item David Gunawan \email{dgunawan@uow.edu.au} (Author of original MATLAB code)
  \item Guy Hawkins \email{guy.hawkins@newcastle.edu.au}
  \item Scott Brown \email{scott.brown@newcastle.edu.au} (Original translation from MATLAB to R) [translator]
}

}
\keyword{internal}
