% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmwg.R, R/tmp.R
\name{is2}
\alias{is2}
\title{Generic function for calculating importance samples}
\usage{
is2(x, n_isamples, n_particles, n_cpus = 1, ...)

is2(x, n_isamples, n_particles, n_cpus = 1, ...)
}
\arguments{
\item{x}{The sampling object you wish to draw importance samples from}

\item{n_isamples}{The number of importance samples to generate}

\item{n_particles}{The number of particles to generate for each iteration of
the is2 process}

\item{n_cpus}{The number of cpus available for parallel processing}

\item{...}{Other arguments passed to methods}
}
\value{
An isample object with the importance samples.

An isample object with the importance samples.
}
\description{
Should be used by creating a S3 method (`is2.object`) for your specific
object.

Should be used by creating a S3 method (`is2.object`) for your specific
object.
}
