% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{mix_gaussian}
\alias{mix_gaussian}
\title{Create a mixture of gaussians for importance sampling}
\usage{
mix_gaussian(parvector, k = 2, maxit = 5000)
}
\arguments{
\item{parvector}{A vector containing}

\item{k}{The number of multivariate gaussians to estimate in the mixture}

\item{maxit}{The maximum number of iterations to pass to `mvnormalmixEM`}
}
\value{
A mix object containing mu and sigma for each gaussian in mix
}
\description{
A slow process, that fails regularly - uses `mixtools::mvnormalmixEM` to
create a mixture of `k` gaussians used in the importance sampling to generate
proposals (particles) from the importance sampling distribution.
}
