% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarise}
\alias{summarise}
\title{Get summary measures of the importance samples}
\usage{
summarise(isamples, se_boot = 10000)
}
\arguments{
\item{isamples}{The importance samples generated by is2}

\item{se_boot}{The number of bootstrap iterations to run}
}
\value{
A list containing the max likelihood and variance using bootstrap
}
\description{
Given the importance samples calculate the maximum likelihood and the
variance using bootstrap procedures
}
