% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.isample}
\alias{summary.isample}
\title{Get summary measures of the importance samples}
\usage{
\method{summary}{isample}(object, se_boot = 10000, ...)
}
\arguments{
\item{object}{The importance samples generated by is2 (isample object)}

\item{se_boot}{The number of bootstrap iterations to run}

\item{...}{Further arguments passed to summary generic function}
}
\value{
A list containing the max likelihood and variance using bootstrap
}
\description{
Given the importance samples calculate the maximum likelihood and the
variance using bootstrap procedures
}
