% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{compute_lw}
\alias{compute_lw}
\title{Calculate importance samples}
\usage{
compute_lw(
  prop_theta,
  data,
  n_subjects,
  n_particles,
  n_randeffect,
  mu_tilde,
  sigma_tilde,
  i,
  prior_dist,
  prior,
  group_dist,
  mix,
  n_params,
  par_names,
  ll_func
)
}
\arguments{
\item{prop_theta}{- the proposals for theta_mu, the group level model
parameter estimates}

\item{data}{- the data used to create the estimates}

\item{n_subjects}{- The number of subjects from the data}

\item{n_particles}{- the number of particles to draw for each importance
sample.}

\item{n_randeffect}{- the number of parameters that has been estimated}

\item{mu_tilde}{- The mean for each subjects random effects}

\item{sigma_tilde}{- The covsriance matrix for each subjects random effects}

\item{i}{- The index for the importance sample being calculated}

\item{prior_dist}{- The specific calculation for the log density of the prior
distrobution}

\item{prior}{- the specification of the prior}

\item{group_dist}{- The specific calculation for the log density for the
group distribution}

\item{mix}{- The calculated values for the importance mixing vars}

\item{n_params}{- The number of parameters in total}

\item{par_names}{- The names of each parameter}

\item{ll_func}{- The log lielihood function passed onto other functions}
}
\value{
The logweight of the importance samples
}
\description{
Calculate importance samples
}
