% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{get_logp}
\alias{get_logp}
\title{Calculate the log probability of the data given the xxx}
\usage{
get_logp(
  prop_theta,
  data,
  n_subjects,
  n_particles,
  n_randeffect,
  mu_tilde,
  sigma_tilde,
  ll_func,
  group_dist,
  n_params,
  par_names
)
}
\arguments{
\item{prop_theta}{- the proposals for theta_mu, the group level model
parameter estimates}

\item{data}{- the data used to create the estimates}

\item{n_subjects}{- The number of subjects from the data}

\item{n_particles}{- the number of particles to draw for each importance
sample.}

\item{n_randeffect}{- the number of parameters that has been estimated}

\item{mu_tilde}{- The mean for each subjects random effects}

\item{sigma_tilde}{- The covsriance matrix for each subjects random effects}

\item{ll_func}{- The log likelihood function to get likelihood of the data
given a set of parameter estimates}

\item{group_dist}{- The specific calculation for the log density for the
group distribution}

\item{n_params}{- The number of parameters in total}

\item{par_names}{- The name of each parameter in the order they appear in the
parameter arrays}
}
\value{
Something or ther - not sure yet.
}
\description{
Given th list of X parameters do some calucltions and return stuff
}
