% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mix.R
\name{get_proposals}
\alias{get_proposals}
\title{Gets proposal mvnormal samples based on mix of gaussians}
\usage{
get_proposals(mix, n_isamples)
}
\arguments{
\item{mix}{A `mixEM` object generated by mixtools via mix_gaussian function.}

\item{n_isamples}{The number of importance samples to generate}
}
\value{
An array with n_isamples proposal importance sample distributions
  drawn from the mixture of gaussians using rmvnorm.
}
\description{
Currently assumes that the mix of gaussians was generated with k = 2
}
