% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmwg.R
\name{is2.pmwgs}
\alias{is2.pmwgs}
\title{S3 method for calculating importance samples for pmwgs objects}
\usage{
\method{is2}{pmwgs}(x, n_isamples, n_particles, ...)
}
\arguments{
\item{x}{The vector of parameter estimates you are working with}

\item{n_isamples}{The number of importance samples to generate}

\item{n_particles}{The number of particles to generate for each iteration of
the is2 process}
}
\value{
An isample object with the importance samples.
}
\description{
Calculates importance samples for a pmwgs object (from the `pmwg` package)
}
